Change: price type from int to float

+++ models/Food.php
@@ -4,7 +4,7 @@ class Food extends PetProduct
 {
     function __construct(
         protected string $name,
-        protected int $price,
+        protected float $price,
         protected string $category,
         private string $type, //dry, weet or snack
         private string $expiration //data
