+++ models/PetProduct.php
@@ -4,7 +4,7 @@ class PetProduct
 {
     function __construct(
         protected string $name,
-        protected int $price,
+        protected float $price,
         protected string $category //dog or cat
     ) {
     }
